+++ documentation/changelog.md
@@ -26,6 +26,8 @@ description: List of all the changes and current versions.
 + Added DLC Clothing Support on Server Side
 + Fix Wanted Level Values and Text
 + Fix Pushing Vehicle Validity
++ Fix Seatbelt Command
++ Add Set Into Vehicle / Vehicle Seat Teleport
 ```
 
 ## 2.0.3